Detect Prod or Test enviornment for Roles and Approvers
Pulled the code that was being used to strip the -TEST off the end of roles
being returned when calling the roles endpoing into the Api base class for better modularity and
reuse.  To detect TEST and PROD environments for the approvers end point
all was that was needed is modify the apps local.php config file on
keyserver-test and prepend "DIPS-TEST" to the approverfilter in the filtergroups array.
The local.php config on keyserver prod did not need to be changed